# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Mock Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  name?: string;
  avatarUrl?: string;
  targetRole?: string;
  experienceLevel?: string;
}

export interface UserStats {
  totalInterviews: number;
  completedInterviews: number;
  averageScore: number;
  currentStreak: number;
  totalHours: number;
  /** @nullable */
  bestScore?: number | null;
  /** @nullable */
  recentImprovement?: number | null;
}

export interface Resume {
  id: number;
  userId: number;
  filename: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  parsedSkills?: string | null;
  isActive?: boolean;
  uploadedAt: string;
}

export interface ResumeInput {
  filename: string;
  content: string;
}

export type InterviewType = typeof InterviewType[keyof typeof InterviewType];


export const InterviewType = {
  technical: 'technical',
  behavioral: 'behavioral',
  system_design: 'system_design',
  mixed: 'mixed',
} as const;

export type InterviewDifficulty = typeof InterviewDifficulty[keyof typeof InterviewDifficulty];


export const InterviewDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type InterviewStatus = typeof InterviewStatus[keyof typeof InterviewStatus];


export const InterviewStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Interview {
  id: number;
  userId: number;
  title: string;
  type: InterviewType;
  difficulty: InterviewDifficulty;
  status: InterviewStatus;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  questionCount?: number;
  answeredCount?: number;
  /** @nullable */
  resumeId?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export type InterviewInputType = typeof InterviewInputType[keyof typeof InterviewInputType];


export const InterviewInputType = {
  technical: 'technical',
  behavioral: 'behavioral',
  system_design: 'system_design',
  mixed: 'mixed',
} as const;

export type InterviewInputDifficulty = typeof InterviewInputDifficulty[keyof typeof InterviewInputDifficulty];


export const InterviewInputDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface InterviewInput {
  title: string;
  type: InterviewInputType;
  difficulty: InterviewInputDifficulty;
  targetRole?: string;
  questionCount?: number;
  resumeId?: number;
}

export interface SkillScore {
  skill: string;
  score: number;
  maxScore: number;
}

export interface InterviewAnalysis {
  interviewId: number;
  overallScore: number;
  summary: string;
  strengths: string[];
  improvements: string[];
  skillScores: SkillScore[];
  /** @nullable */
  detailedFeedback?: string | null;
}

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  technical: 'technical',
  behavioral: 'behavioral',
  situational: 'situational',
  coding: 'coding',
} as const;

export interface Question {
  id: number;
  interviewId: number;
  text: string;
  type: QuestionType;
  orderIndex: number;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  userAnswer?: string | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  feedback?: string | null;
  isAnswered?: boolean;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
}

export interface AnswerResult {
  questionId: number;
  score: number;
  feedback: string;
  strengths?: string[];
  improvements?: string[];
}

export interface AnalyticsOverview {
  totalInterviews: number;
  averageScore: number;
  completionRate: number;
  topSkill: string;
  weakestSkill: string;
  /** @nullable */
  scoreThisWeek?: number | null;
  /** @nullable */
  scoreLastWeek?: number | null;
}

export interface ScorePoint {
  date: string;
  score: number;
  /** @nullable */
  interviewId?: number | null;
}

export interface CategoryScore {
  category: string;
  averageScore: number;
  count: number;
}

export type ListInterviewsParams = {
status?: ListInterviewsStatus;
};

export type ListInterviewsStatus = typeof ListInterviewsStatus[keyof typeof ListInterviewsStatus];


export const ListInterviewsStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type GetScoreTrendParams = {
days?: number;
};